Replace float

+++ schema.sql
@@ -6,6 +6,6 @@ CREATE TABLE animals (
     date_of_birth date,
     escape_attempts int,
     neutred boolean,
-    weight_kg float
+    weight_kg DECIMAL(10,2)
 );
